Restructure app as a package
