+++ app/__init__.py
@@ -1,0 +1,4 @@
+# Set default log format
+import logging
+LOG_FORMAT = '%(name)-12s %(levelname)-8s %(message)s'
+logging.basicConfig(format=LOG_FORMAT)
