+++ app/services/contract_observer.py
@@ -1,19 +1,14 @@
-#!/usr/bin/env python
-import sys
-sys.path.insert(0, '/usr/src/app')
-sys.path.insert(0, '/usr/src/app/src')
-
-from app import App
+from ..app import App
 import asyncio
-from config import ED_CONTRACT_ADDR, ED_CONTRACT_ABI, WS_PROVIDER_URL
-from contract_event_recorders import record_cancel, record_deposit, process_trade, record_withdraw
-from contract_event_utils import block_timestamp
+from ..config import ED_CONTRACT_ADDR, ED_CONTRACT_ABI, WS_PROVIDER_URL
+from ..src.contract_event_recorders import record_cancel, record_deposit, process_trade, record_withdraw
+from ..src.contract_event_utils import block_timestamp
 import json
 import logging
 from time import time
-from utils import coerce_to_int
+from ..src.utils import coerce_to_int
 from websockets import connect
-from websocket_filter_set import WebsocketFilterSet
+from ..src.websocket_filter_set import WebsocketFilterSet
 
 logger = logging.getLogger("contract_observer")
