+++ app/src/contract_event_recorders.py
@@ -1,11 +1,11 @@
-from app import App
-from contract_event_utils import block_timestamp
+from ..app import App
+from app.src.contract_event_utils import block_timestamp
 from datetime import datetime
 import logging
-from order_enums import OrderSource, OrderState
-from order_hash import make_order_hash
+from app.src.order_enums import OrderSource, OrderState
+from app.src.order_hash import make_order_hash
 from pprint import pprint
-from utils import coerce_to_int, parse_insert_status
+from app.src.utils import coerce_to_int, parse_insert_status
 from web3 import Web3
 
 ZERO_ADDR = "0x0000000000000000000000000000000000000000"
